requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-paypal==0.5.0
idna==2.7
lxml==4.2.5
oauthlib==2.1.0
python3-memcached==1.51
python3-openid==3.1.0
pytz==2018.7
requests==2.20.0
requests-oauthlib==1.0.0
selenium==3.141.0
six==1.11.0
urllib3==1.24.1
